# openapi.json
{"openapi":"3.1.0","info":{"title":"VBoarder API","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","description":"Welcome endpoint.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/chat/{agent_role}":{"post":{"summary":"Chat With Agent","description":"Chat with a specific C-suite agent with session management.\n\nArgs:\n    agent_role: The role (air, ceo, cfo, clo, cmo, coo, cos, cto, sec)\n    request: ChatRequest with message and optional session_id\n\nReturns:\n    Agent response with session context","operationId":"chat_with_agent_chat__agent_role__post","parameters":[{"name":"agent_role","in":"path","required":true,"schema":{"type":"string","title":"Agent Role"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sessions/{agent_role}":{"get":{"summary":"List Sessions","description":"List all sessions for a specific agent.","operationId":"list_sessions_sessions__agent_role__get","parameters":[{"name":"agent_role","in":"path","required":true,"schema":{"type":"string","title":"Agent Role"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sessions":{"get":{"summary":"List All Sessions","description":"List all sessions across all agents.\nUseful for debugging and monitoring.","operationId":"list_all_sessions_sessions_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/sessions/{agent_role}/{session_id}":{"delete":{"summary":"Delete Session","description":"Delete a specific session for an agent.","operationId":"delete_session_sessions__agent_role___session_id__delete","parameters":[{"name":"agent_role","in":"path","required":true,"schema":{"type":"string","title":"Agent Role"}},{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/agents":{"get":{"summary":"List Agents","description":"List all available agents with their configurations.","operationId":"list_agents_agents_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ChatRequest":{"properties":{"message":{"type":"string","title":"Message"},"session_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Session Id","default":"default"},"concise":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Concise","default":false}},"type":"object","required":["message"],"title":"ChatRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}